+++ pdf/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Improve Bullet Widget
 - Use covariant on SpanningWidget
 - ImageProvider.resolve returns non-null object
+- Fix textScalingFactor with lineSpacing
 
 ## 3.2.0